order page terminada dire
ver si agregar la dirección
sino se hace centrar el nombre de la dirección

+++ Irui2/app/src/main/java/com/example/usuario/irui/requestModels/Order.java
@@ -107,9 +107,9 @@ public class Order{
         return longitude;
     }
 
-    public int getSubtotal() {
+    public double getSubtotal() {
 
-        int subtotal = 0;
+        double subtotal = 0;
 
         for (int i = 0; i < items.length; i++) {
             Item item = items[i];
